moved utils.py to src and updates imports
